refactor: change pages, constants folder name to camelCase
