+++ src/pages/authPage/AuthPage.styles.ts
